hackaton1/requirements.txt
aiogram
python-dotenv
Pillow
requests
replicate